Address Copilot PR review comments (#1290)
- Fix double GetAdviceOrderIndices call in IntroduceExtensionBlockAdvice
- Improve OperatorHasWrongStaticity diagnostic message (use "static"/"non-static")
- Update IExtensionBlockBuilder docs for Name property
- Update test expected outputs for new ordering
