+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Introductions/ExtensionBlocks/MultipleBlocksDifferentTypes.t.cs
@@ -1,10 +1,10 @@
 [IntroductionAttribute]
 public static class TargetType
 {
-  extension(global::System.Int32 i)
+  extension(global::System.String s)
   {
   }
-  extension(global::System.String s)
+  extension(global::System.Int32 i)
   {
   }
 }
