+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Introductions/ExtensionBlocks/MultipleBlocksSameType.t.cs
@@ -1,10 +1,10 @@
 [IntroductionAttribute]
 public static class TargetType
 {
-  extension(global::System.String self2)
+  extension(global::System.String self1)
   {
   }
-  extension(global::System.String self1)
+  extension(global::System.String self2)
   {
   }
 }
